// src/api/models/ApartmentsFilters.ts
export interface ApartmentsFilters {
  /**
   * Адрес. Ищет плохо, вместо адреса стоит лучше использовать toponym
   * @type {string}
   * @memberof ApartmentsFilters
   */
  address?: string
  /**
   *
   * @type {string}
   * @memberof ApartmentsFilters
   */
  city?: string | null
  /**
   * Топоним. Формат Яндекс гео. Если передать топоним, то фильтр Адрес игнорируется
   * @type {Array<ApartmentsFiltersToponym>}
   * @memberof ApartmentsFilters
   */
  toponym?: Array<ApartmentsFiltersToponym>
  /**
   *
   * @type {Array<string>}
   * @memberof ApartmentsFilters
   */
  roominess?: Array<string>
  /**
   *
   * @type {number}
   * @memberof ApartmentsFilters
   */
  priceRubMin?: number
  /**
   *
   * @type {number}
   * @memberof ApartmentsFilters
   */
  priceRubMax?: number
  /**
   *
   * @type {number}
   * @memberof ApartmentsFilters
   */
  fullAreaMin?: number
  /**
   *
   * @type {number}
   * @memberof ApartmentsFilters
   */
  fullAreaMax?: number
  /**
   *
   * @type {number}
   * @memberof ApartmentsFilters
   */
  kitchensAreaMin?: number
  /**
   *
   * @type {number}
   * @memberof ApartmentsFilters
   */
  kitchensAreaMax?: number
  /**
   *
   * @type {number}
   * @memberof ApartmentsFilters
   */
  livingAreaMin?: number
  /**
   *
   * @type {number}
   * @memberof ApartmentsFilters
   */
  livingAreaMax?: number
  /**
   *
   * @type {number}
   * @memberof ApartmentsFilters
   */
  buildingFloorCountMin?: number
  /**
   *
   * @type {number}
   * @memberof ApartmentsFilters
   */
  buildingFloorCountMax?: number
  /**
   *
   * @type {number}
   * @memberof ApartmentsFilters
   */
  floorNumberMin?: number
  /**
   *
   * @type {number}
   * @memberof ApartmentsFilters
   */
  floorNumberMax?: number
  /**
   *
   * @type {Array<string>}
   * @memberof ApartmentsFilters
   */
  buildingCeilingHeight?: Array<ApartmentsFiltersBuildingCeilingHeightEnum>
  /**
   *
   * @type {Array<string>}
   * @memberof ApartmentsFilters
   */
  housingTypes?: Array<string>
  /**
   *
   * @type {Array<string>}
   * @memberof ApartmentsFilters
   */
  saleTypes?: Array<string>
  /**
   *
   * @type {Array<string>}
   * @memberof ApartmentsFilters
   */
  decorationTypes?: Array<string>
  /**
   *
   * @type {Array<string>}
   * @memberof ApartmentsFilters
   */
  buildingTypes?: Array<string>
  /**
   *
   * @type {Array<string>}
   * @memberof ApartmentsFilters
   */
  windowViewTypes?: Array<string>
  /**
   *
   * @type {Array<string>}
   * @memberof ApartmentsFilters
   */
  buildingParkingTypes?: Array<string>
  /**
   *
   * @type {Array<string>}
   * @memberof ApartmentsFilters
   */
  wcsType?: Array<ApartmentsFiltersWcsTypeEnum>
  /**
   *
   * @type {Array<string>}
   * @memberof ApartmentsFilters
   */
  elevatorType?: Array<ApartmentsFiltersElevatorTypeEnum>
  /**
   *
   * @type {Array<string>}
   * @memberof ApartmentsFilters
   */
  balconyType?: Array<ApartmentsFiltersBalconyTypeEnum>
  /**
   *
   * @type {Array<string>}
   * @memberof ApartmentsFilters
   */
  buildingGarbageChute?: Array<ApartmentsFiltersBuildingGarbageChuteEnum>
  /**
   * Фильтр по id квартир
   * @type {Array<number>}
   * @memberof ApartmentsFilters
   */
  apartments?: Array<number>
  /**
   *
   * @type {number}
   * @memberof ApartmentsFilters
   */
  loanRateMin?: number
  /**
   *
   * @type {number}
   * @memberof ApartmentsFilters
   */
  loanRateMax?: number
  /**
   *
   * @type {ApartmentsFiltersBoundingBox}
   * @memberof ApartmentsFilters
   */
  boundingBox?: ApartmentsFiltersBoundingBox
}

/**
 * @export
 * @enum {string}
 */
export enum ApartmentsFiltersBuildingCeilingHeightEnum {
  NoMatter = 'no-matter',
  _250 = '>=250',
  _270 = '>=270',
  _300 = '>=300',
  _350 = '>=350',
  _400 = '>=400',
}
/**
 * @export
 * @enum {string}
 */
export enum ApartmentsFiltersWcsTypeEnum {
  NoMatter = 'no-matter',
  Combined = 'combined',
  Separate = 'separate',
}
/**
 * @export
 * @enum {string}
 */
export enum ApartmentsFiltersElevatorTypeEnum {
  NoMatter = 'no-matter',
  Any = 'any',
  Freight = 'freight',
}
/**
 * @export
 * @enum {string}
 */
export enum ApartmentsFiltersBalconyTypeEnum {
  NoMatter = 'no-matter',
  Loggia = 'loggia',
  Balcony = 'balcony',
}
/**
 * @export
 * @enum {string}
 */
export enum ApartmentsFiltersBuildingGarbageChuteEnum {
  NoMatter = 'no-matter',
  Yes = 'yes',
  No = 'no',
}